2469: Expect RecordNotFound instead of 404
Setting config.action_dispatch.show_exceptions to false disables the part of
Rails that turns an ActiveRecord::RecordNotFound into a 404 status code.

At some point, it may be worthwile to add a rescue_from
ActiveRecord::RecordNotFound in the base controller to explicitly return a 404
ourselves, perhaps only in the spec_helper.