2312: Remove unnecessary queries for SMS creation.
The user was already retrieved from the database for the message,
so no need to get it again on Decode.
Changed adapters to return the message instance and the SMS' controller
is responsible to set the mission. Only then it creates the message, thus
avoiding the two queries (one to create and another to update with the
mission).