3074: Improve non observers show on StandardFormReport.
The system was fetching all users to show them on the form. This was
making the system extremely slow when it had a lot of users on the
database. I created one query to fetch the users limiting how many
we are getting and used SQL_CALC_FOUND_ROWS to get the total count
without hitting the db again.
Also added a composite index to speed the query on responses because
it was slow on a system with a lot of responses.