8636: Shift index links to group batch links and collection links
The ELMO theme accent color was changed from orange to blue so that we wouldn't need to use the $darker-accent-color in index tables, which was kind of a hack. Orange as the link color, which Boostrap sets as $accent-color, was always a mistake. $darker-accent-color is now used for collection index links while the default $accent-color is used for batch index links.